requirements.txt
pytest~=6.1.2
Appium-Python-Client==1.0.2
selenium>=3.141.0
requests>=2.25.1